% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{autocomplete_score}
\alias{autocomplete_score}
\title{Get the autocomplete from nav system.}
\usage{
autocomplete_score(nav_system)
}
\arguments{
\item{commands}{Vector of nav system entries}
}
\value{
The autocomplete score
}
\description{
Get the autocomplete from nav system.
}
