% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{basin_sizes}
\alias{basin_sizes}
\title{Get the value of largest 3 basins multiplied together}
\usage{
basin_sizes(map)
}
\arguments{
\item{map}{height map}
}
\value{
The size of 3 largest basins multiplied together
}
\description{
Get the value of largest 3 basins multiplied together
}
