% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_dangerous_areas_2}
\alias{count_dangerous_areas_2}
\title{Count dangerous areas from input string of vent paths}
\usage{
count_dangerous_areas_2(vents)
}
\arguments{
\item{vents}{Vent paths}
}
\value{
The count of dangerous areas
}
\description{
Count dangerous areas from input string of vent paths
}
