% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_digits}
\alias{count_digits}
\title{Count the number of times 1, 4, 7 or 8 appear in log entries}
\usage{
count_digits(entries)
}
\arguments{
\item{entries}{Log entries of form 10 signal patterns | delimitter and
4 output values}
}
\value{
The number of times 1, 4, 7 or 8 appear
}
\description{
Count the number of times 1, 4, 7 or 8 appear in log entries
}
