% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_fish}
\alias{count_fish}
\title{Count the number of fish after a number of days from some internal state}
\usage{
count_fish(fish, days)
}
\arguments{
\item{fish}{Vector of fish internal timers}

\item{days}{The number of days to run}
}
\value{
The final number of fish
}
\description{
Count the number of fish after a number of days from some internal state
}
