% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_increases}
\alias{count_increases}
\title{Count number of times a sequence of numbers increases.}
\usage{
count_increases(values)
}
\arguments{
\item{values}{Vector of integers}
}
\value{
The number of times subsequent value increases
}
\description{
Count number of times a sequence of numbers increases.
}
