% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{error_score}
\alias{error_score}
\title{Get the error score from nav system.}
\usage{
error_score(nav_system)
}
\arguments{
\item{commands}{Vector of nav system entries}
}
\value{
The syntax error score
}
\description{
Get the error score from nav system.
}
