% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Get the coordinates x1, y1, x2, y2 from input string of vent}
\usage{
get_coordinates(vents)
}
\arguments{
\item{commands}{Vector of commands for the submarine,
e.g. forward 2, down 3 or up 5}
}
\value{
The distance traveled by the submarine, resultant depth * horizontal
}
\description{
Takes vector of strings of format "0,9 -> 5,9" and parses to vector
of arrays
}
