% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_first_all_flash}
\alias{get_first_all_flash}
\title{Get the first day all octopuses flash}
\usage{
get_first_all_flash(octopuses, steps)
}
\arguments{
\item{octopuses}{Grid of octopuses}

\item{steps}{The number of steps to run}
}
\value{
The first day all octopuses flash
}
\description{
Get the first day all octopuses flash
}
