% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_flashes}
\alias{get_flashes}
\title{Count the number of times grid of octopuses flashed}
\usage{
get_flashes(octopuses, steps)
}
\arguments{
\item{octopuses}{Grid of octopuses}

\item{steps}{The number of steps to run}
}
\value{
The total number of times an octopus flashed
}
\description{
Count the number of times grid of octopuses flashed
}
