% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_low_points}
\alias{get_low_points}
\title{Get the low points from heightmap}
\usage{
get_low_points(parsed)
}
\arguments{
\item{map}{Heightmap}
}
\value{
The low points in the heightmap
}
\description{
Get the low points from heightmap
}
