% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_min_total_risk}
\alias{get_min_total_risk}
\title{Get the min risk of traversing full risk grid}
\usage{
get_min_total_risk(risk)
}
\arguments{
\item{risk}{Vector of vector of ints of risk of moving to a space}
}
\value{
The minimum total risk
}
\description{
Get the min risk of traversing full risk grid
}
