% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_paths}
\alias{get_num_paths}
\title{Get the number of complete paths through caves}
\usage{
get_num_paths(connections, small_caves, verbose)
}
\arguments{
\item{small_caves}{The number of times you can visit a small cave}

\item{verbose}{If TRUE print the completed paths}

\item{commands}{Vector of size 2 vectors, showing connections from caves}
}
\value{
The total number of complete caves
}
\description{
Get the number of complete paths through caves
}
