% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_package_value}
\alias{get_package_value}
\title{Get the value of package}
\usage{
get_package_value(input)
}
\arguments{
\item{input}{The packet hexadecimal}
}
\value{
The sum of version numbers
}
\description{
Get the value of package
}
