% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{life_support_rating}
\alias{life_support_rating}
\title{Get the life support rating from set of bytes}
\usage{
life_support_rating(numbers)
}
\arguments{
\item{numbers}{Vector of binary numbers}
}
\value{
The life support rating
}
\description{
Get the life support rating from set of bytes
}
