% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{model_fish}
\alias{model_fish}
\title{Model fish internal timers for a number of days for some initial fish state.}
\usage{
model_fish(fish, days)
}
\arguments{
\item{fish}{Vector of fish internal timers}

\item{days}{The number of days to run}
}
\value{
The final state of fish internal timers
}
\description{
Model fish internal timers for a number of days for some initial fish state.
}
