% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimal_position}
\alias{optimal_position}
\title{Get the lowest amount of fuel needed for all crabs to align.}
\usage{
optimal_position(positions)
}
\arguments{
\item{positions}{Vector of crab horizontal positions}
}
\value{
The minimum fuel needed for the crabs to align
}
\description{
Get the lowest amount of fuel needed for all crabs to align.
}
