% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimal_position_2}
\alias{optimal_position_2}
\title{Get the lowest amount of fuel needed for all crabs to align puzzle 2.}
\usage{
optimal_position_2(positions)
}
\arguments{
\item{positions}{Vector of crab horizontal positions}
}
\value{
The minimum fuel needed for the crabs to align
}
\description{
Get the lowest amount of fuel needed for all crabs to align puzzle 2.
}
