% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polymer_size}
\alias{polymer_size}
\title{Print the diff in size between most and least common element}
\usage{
polymer_size(input, times)
}
\arguments{
\item{input}{Puzzle input}
}
\description{
Print the diff in size between most and least common element
}
