% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{position_cost_2}
\alias{position_cost_2}
\title{Get the cost of aligning crabs to a particular position}
\usage{
position_cost_2(positions, alignment_position)
}
\arguments{
\item{positions}{Vector of crab positions}

\item{alignment_position}{Final crab positions}
}
\value{
The fuel cost of aligning to that position
}
\description{
Get the cost of aligning crabs to a particular position
}
