% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{power_consumption}
\alias{power_consumption}
\title{Get the power consumption from set of bytes}
\usage{
power_consumption(numbers)
}
\arguments{
\item{numbers}{Vector of binary numbers}
}
\value{
The power consumption
}
\description{
Get the power consumption from set of bytes
}
