% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{risk_level}
\alias{risk_level}
\title{Get the risk level from heightmap}
\usage{
risk_level(map)
}
\arguments{
\item{map}{Heightmap}
}
\value{
The risk level
}
\description{
Get the risk level from heightmap
}
