% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_day}
\alias{run_day}
\title{Run a single day through fish timer.}
\usage{
run_day(fish)
}
\arguments{
\item{fish}{Vector of fish internal timers}
}
\value{
The final state of fish internal timers
}
\description{
Take a vector of fish timers, reduce their values by 1 or if 0
reset timer to 6. If timer was reset, add a new fish to the end
with start timer of 8.
}
