% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_10.R
\name{solve_day_10}
\alias{solve_day_10}
\title{Solve day 10 problem 1}
\usage{
solve_day_10(file)
}
\arguments{
\item{file}{Path to file of syntax errors}
}
\value{
The syntax error score
}
\description{
Solve day 10 problem 1
}
