% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_10.R
\name{solve_day_10_puzzle_2}
\alias{solve_day_10_puzzle_2}
\title{Solve day 10 puzzle 2}
\usage{
solve_day_10_puzzle_2(file)
}
\arguments{
\item{file}{Path to file of syntax errors}
}
\value{
The autocomplete score
}
\description{
Solve day 10 puzzle 2
}
