% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_11.R
\name{solve_day_11}
\alias{solve_day_11}
\title{Solve day 11 problem 1}
\usage{
solve_day_11(file)
}
\arguments{
\item{file}{Path to file of syntax errors}
}
\value{
The syntax error score
}
\description{
Solve day 11 problem 1
}
