% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_11.R
\name{solve_day_11_puzzle_2}
\alias{solve_day_11_puzzle_2}
\title{Solve day 11 problem 2}
\usage{
solve_day_11_puzzle_2(file)
}
\arguments{
\item{file}{Path to file of syntax errors}
}
\value{
The syntax error score
}
\description{
Solve day 11 problem 2
}
