% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_12.R
\name{solve_day_12}
\alias{solve_day_12}
\title{Solve day 12 problem 1}
\usage{
solve_day_12(file)
}
\arguments{
\item{file}{Path to file of cave connections}
}
\value{
The total number of possible paths
}
\description{
Solve day 12 problem 1
}
