% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_13.R
\name{solve_day_13_puzzle_2}
\alias{solve_day_13_puzzle_2}
\title{Solve day 13 problem 2}
\usage{
solve_day_13_puzzle_2(file)
}
\arguments{
\item{file}{Path to input}
}
\value{
Prints the answer
}
\description{
Solve day 13 problem 2
}
