% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_14.R
\name{solve_day_14_puzzle_2}
\alias{solve_day_14_puzzle_2}
\title{Solve day 14 problem 2}
\usage{
solve_day_14_puzzle_2(file)
}
\arguments{
\item{file}{Path to input}
}
\value{
The diff
}
\description{
Solve day 14 problem 2
}
