% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_15.R
\name{solve_day_15_puzzle_2}
\alias{solve_day_15_puzzle_2}
\title{Solve day 15 problem 2}
\usage{
solve_day_15_puzzle_2(file)
}
\arguments{
\item{file}{Path to input}
}
\value{
The minimum risk
}
\description{
Solve day 15 problem 2
}
