% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_16.R
\name{solve_day_16}
\alias{solve_day_16}
\title{Solve day 16 problem 1}
\usage{
solve_day_16(file)
}
\arguments{
\item{file}{Path to input}
}
\value{
The sum of version numbers
}
\description{
Solve day 16 problem 1
}
