% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_16.R
\name{solve_day_16_puzzle_2}
\alias{solve_day_16_puzzle_2}
\title{Solve day 16 problem 2}
\usage{
solve_day_16_puzzle_2(file)
}
\arguments{
\item{file}{Path to input}
}
\value{
The value of package
}
\description{
Solve day 16 problem 2
}
