% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_2.R
\name{solve_day_2}
\alias{solve_day_2}
\title{Solve day 2 problem 1}
\usage{
solve_day_2(file)
}
\arguments{
\item{file}{Path to file of commands}
}
\value{
The number of times a depth measurement increases
}
\description{
Solve day 2 problem 1
}
