% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_3.R
\name{solve_day_3}
\alias{solve_day_3}
\title{Solve day 3 problem 1}
\usage{
solve_day_3(file)
}
\arguments{
\item{file}{Path to file of numbers}
}
\value{
The power consumption of the submarine
}
\description{
Solve day 3 problem 1
}
