% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_3.R
\name{solve_day_3_puzzle_2}
\alias{solve_day_3_puzzle_2}
\title{Solve day 3 problem 2}
\usage{
solve_day_3_puzzle_2(file)
}
\arguments{
\item{file}{Path to file of numbers}
}
\value{
The life support rating of the submarine
}
\description{
Solve day 3 problem 2
}
