% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_4.R
\name{solve_day_4}
\alias{solve_day_4}
\title{Solve day 4 problem 1}
\usage{
solve_day_4(file)
}
\arguments{
\item{file}{Path to file of numbers}
}
\value{
The power consumption of the submarine
}
\description{
Solve day 4 problem 1
}
