% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_5.R
\name{solve_day_5}
\alias{solve_day_5}
\title{Solve day 5 problem 1}
\usage{
solve_day_5(file)
}
\arguments{
\item{file}{Path to file of vents}
}
\value{
The number of dangerous areas i.e. 2 or more vents
}
\description{
Solve day 5 problem 1
}
