% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_6.R
\name{solve_day_6}
\alias{solve_day_6}
\title{Solve day 6 problem 1}
\usage{
solve_day_6(file)
}
\arguments{
\item{file}{Path to file of depth measurements}
}
\value{
The number of times a depth measurement increases
}
\description{
Solve day 6 problem 1
}
