% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_6.R
\name{solve_day_6_puzzle_2}
\alias{solve_day_6_puzzle_2}
\title{Solve day 6 problem 2}
\usage{
solve_day_6_puzzle_2(file)
}
\arguments{
\item{file}{Path to file of depth measurements}
}
\value{
The number of times a depth measurement increases
}
\description{
Solve day 6 problem 2
}
