% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_7.R
\name{solve_day_7}
\alias{solve_day_7}
\title{Solve day 7 problem 1}
\usage{
solve_day_7(file)
}
\arguments{
\item{file}{Path to file of crab positions}
}
\value{
The minimum fuel needed to align the crabs
}
\description{
Solve day 7 problem 1
}
