% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_8.R
\name{solve_day_8_puzzle_2}
\alias{solve_day_8_puzzle_2}
\title{Solve day 8 problem 2}
\usage{
solve_day_8_puzzle_2(file)
}
\arguments{
\item{file}{Path to file of entries}
}
\value{
The count of digits 1, 4, 7 and 8 in outputs
}
\description{
Solve day 8 problem 2
}
