% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_9.R
\name{solve_day_9_puzzle_2}
\alias{solve_day_9_puzzle_2}
\title{Solve day 9 problem 2}
\usage{
solve_day_9_puzzle_2(file)
}
\arguments{
\item{file}{Path to file height map}
}
\value{
The 3 largest basin sizes multiplied
}
\description{
Solve day 9 problem 2
}
