% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_diff}
\alias{str_diff}
\title{Remove characters of b from a.}
\usage{
str_diff(a, b)
}
\arguments{
\item{a}{string to remove from}

\item{b}{the string to remove}
}
\value{
String a with characters from b removed
}
\description{
Remove characters of b from a.
}
