% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_in}
\alias{str_in}
\title{Check if string b is in a, in any order.}
\usage{
str_in(a, b)
}
\arguments{
\item{a}{superstring}

\item{b}{substring to test.}
}
\value{
true if all characters in b are in a
}
\description{
Check if string b is in a, in any order.
}
