% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_version_numbers}
\alias{sum_version_numbers}
\title{Sum version numbers of all packets}
\usage{
sum_version_numbers(input)
}
\arguments{
\item{input}{The packet hexadecimal}
}
\value{
The sum of version numbers
}
\description{
Sum version numbers of all packets
}
