% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{total_distance_2}
\alias{total_distance_2}
\title{Get the distance traveled from a set of commands for puzzle 2.}
\usage{
total_distance_2(commands)
}
\arguments{
\item{commands}{Vector of commands for the submarine,
e.g. forward 2, down 3 or up 5}
}
\value{
The distance traveled by the submarine, resultant depth * horizontal
}
\description{
This is the result horizontal * result depth.
}
