% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{total_dots}
\alias{total_dots}
\title{Get the total number of marked dots after folding}
\usage{
total_dots(input)
}
\arguments{
\item{input}{Puzzle input}
}
\value{
Number of marked dots
}
\description{
Get the total number of marked dots after folding
}
