% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{window_increases}
\alias{window_increases}
\title{Count number of times a three-measurement sliding window increases in value.}
\usage{
window_increases(values)
}
\arguments{
\item{values}{Vector of integers}
}
\value{
The number of times three measurement sliding window increases.
}
\description{
Count number of times a three-measurement sliding window increases in value.
}
