% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day1.R
\name{count_increases}
\alias{count_increases}
\title{Count number of increases}
\usage{
count_increases(data)
}
\arguments{
\item{data}{data with depth and increase flagged}
}
\value{

}
\description{
Count number of increases
}
