% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_1.R
\name{solve_day_1_puzzle_2}
\alias{solve_day_1_puzzle_2}
\title{Solve day 1 problem 2}
\usage{
solve_day_1_puzzle_2(file)
}
\arguments{
\item{file}{Path to file of component masses}
}
\value{
The fuel required to launch all the components and the fuel itself
}
\description{
Solve day 1 problem 2
}
